refactor: move CalendarScreen style into own file
